ANTARCTICLE-247 Case-insensitive authentication in Antarcticle
Use case:
there is a user with "Petya" username in antarcticle database
there is no user with "petya" username in antarcticle database
there is a user with "patya" username in poulpe database
User "petya" tries to login on

When user was authenticated against poulpe the new record should be added to users database (if there is no such user in antarcticle database) or record with exact (case-sensitive) username should be updated (if it is already persists in antarcticle database).

But it updated the first record ("Petya") as case-insensitive selection was made before insert/update to antarcticle database.