Merge branch 'ANTARCTICLE-119'
Added language and sourceId columns to articles table. Column sourceId is used to indicate the id of primary (source) article that was created for the first time (not translated). All other article translated based on this source article including the primary one have the same sourceId value that equals to the id of the source article.

 Added ability to save translation.